Refactor transform Set out of Face
Currently Set itself is still defined in `objects/full/face.rs` didn't
know where to move it...
