+++ crates/fj-kernel/src/algorithms/transform/face.rs
@@ -1,10 +1,6 @@
 use fj_math::Transform;
 
-use crate::{
-    objects::{Face, Set},
-    operations::Insert,
-    services::Services,
-};
+use crate::{objects::Face, services::Services};
 
 use super::{TransformCache, TransformObject};
 
@@ -31,22 +27,5 @@ impl TransformObject for Face {
         });
 
         Self::new(surface, exterior, interiors, color)
-    }
-}
-
-impl<T: Ord + TransformObject + Clone + Insert + 'static> TransformObject
-    for Set<T>
-{
-    fn transform_with_cache(
-        self,
-        transform: &Transform,
-        services: &mut Services,
-        cache: &mut TransformCache,
-    ) -> Self {
-        let mut inner = Self::new();
-        inner.extend(self.into_iter().map(|i| {
-            TransformObject::transform_with_cache(i, transform, services, cache)
-        }));
-        inner
     }
 }
